+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - limit the rate of MDN sending #3402
 - ignore ratelimits for bots #3439
 - remove `msgs_mdns` references to deleted messages during housekeeping #3387
+- format message lines starting with `>` as quotes #3434
 
 ### Fixes
 - set a default error if NDN does not provide an error